added translations, added checkbox group heading
